+++ src/lib/gui/text.cpp
@@ -14,8 +14,8 @@ Text *Text::draw(Ctx *ctx)
 	//allready locked by GUI thread
 	auto font = get_string_prop("font");
 	auto text = get_string_prop("text");
-	auto col = get_long_prop("color");
-	auto ink_col = get_long_prop("ink_color");
+	auto col = (uint32_t)get_long_prop("color");
+	auto ink_col = (uint32_t)get_long_prop("ink_color");
 	ctx->set_color(col);
 	ctx->filled_box(0, 0, m_w, m_h);
 	ctx->set_color(ink_col);
